should make look nicer, quick maths

+++ src/config.h
@@ -4,8 +4,8 @@
 
 namespace plane {
 struct Config {
-  uint64_t screen_width;
-  uint64_t screen_height;
+  double screen_width;
+  double screen_height;
 };
 
 static Config config = { 640, 480 };
